Add serializers for *Timedelta types
Add serializers to `HumanTimedelta` and `SecondsTimedelta` that
serialize those Pydantic fields to a float number of seconds instead
of ISO 8601 durations. This means those data types now can be
round-tripped (serialized and then deserialized to the original
value), whereas before they could not be. This also avoids ISO 8601
durations in service replies that use models including those types.

Fix parsing of stringified floating-point number of seconds for
`SecondsTimedelta`, which previously was truncating to an integer
number of seconds.