Fix IVOA DALI timestamp parsing, add data types
Change `parse_isodatetime` and `normalize_isodatetime` to support
exactly the formats specified by the IVOA DALI standard (and fix
incorrect references to UWS). This means supporting timestamps
without the time component and making the trailing Z optional, but
requiring seconds be present if the time portion is present. This
should not interfere with the other uses, such as parsing Kubernetes
timestamps.

Add new data types `UtcDatetime` and `IvoaIsoDatetime` that are now
the preferred way to use the normalization functions in models. The
former accepts any format Pydantic supports but coerces it to UTC
and timezone-aware. The latter accepts only the formats supported by
`parse_isodatetime`. Document those data types instead of the
normalization functions, although the normalization functions are
still available (for now).