Add generic paginated query support
Lift the pagninated query support from Gafaelfawr into Safir and
make it generic so that it can be reused in Wobbly.

Pagination support consists of a cursor class, which defines how
results are ordered and paginated and how the cursor is serialized
and deserialized, a paginated query runner that contains the logic
for applying limits and keyset pagination, and a paginated result
container that wraps a list of Pydantic models with metadata about
the pagination. Also define a cursor type for the common case of
results ordered by a `datetime` and then a unique key, in descending
order, so that cursors for that case can be easily constructed.

Fix the documentation build with current Sphinx. Adding `TypeVar`
variables to the exports for documentation purposes is apparently
now pointless since the documentation string appears to be forced
to be generic documentation about type variables, so stop trying to
include them.