Improve pagination documentation
Document that the `id_column` of a `DatetimeIdCursor` must be an
integer. Document in multiple places that forward cursors include
the entry they are based on but previous cursors do not and should
return data starting with the entry immediately previous. Document
that the query is run twice, once with restrictions and then again
with `COUNT` to get the total number of available entries, and
therefore the `COUNT` query needs to be fast (and thus based on
indices or short tables).