Convert the UWS library to use the Wobbly backend
Rather than storing UWS jobs directly in a database, which requires
every UWS-based application to manage its own separate database, use
the Wobbly service to manage all job storage. This service uses a
delegated token to determine the user and service, so considerably
less tracking of the user is required.

UWS applications now store the serialized parameter model in the
database rather than a list of key/value pairs, and rely on methods
on the parameters model to convert to the XML format for the current
IVOA UWS protocol.

Add a mock for Wobbly that can be used to test UWS applications
without having the Wobbly API available.

Drop the `ErrorCode` enum, since its values were specific to SODA,
and instead take the error code as a string. Drop some related
exceptions that are not used directly in Safir and are specific to
SODA.