Add schemas for most UWS routes
Provide a schema for the routes returning vo-models Pydantic XML
models. These may not be entirely correct for the XML schema, but
they seem better than nothing. Do not return a schema for the
`/parameters` route, since we don't have easy access to the correct
parameters model in the handler.