Fix dependency constraints on Safir modules
The main safir PyPI module was constraining the major version of
safir-logging and safir-arq, which would break the new UWS support
by pinning too old of a version of safir-arq. Those major version
constraints are too easy to forget to update, so remove them and
instead have Safir pull in the latest version.

Unfortunately, either approach here potentially creates problems
down the road. Hopefully this will be the approach that creates
fewer issues.