Update metrics configuration documentation
Update the documentation for how to set up metrics events to use
`metrics_configuration_factory` and do not bother to set
`METRICS_EVENTS_TOPIC_PREFIX` in the mock case (it's not needed).
Move `model_config` to the top of the class to make it somewhat
more visible, following the coding style I've been trying to adopt
elsewhere.