Simplify if underlying exception has no message
In `SlackWebException.from_exception`, the message of the created
`SlackWebException` is based on the type name and message of the
underlying exception. However, some underlying exceptions, such as
timeout errors, may have no message. In that case, drop the trailing
colon and space and set the message of `SlackWebException` to only
the name of the type of the underlying exception.