Add object creation callbacks to Kubernetes mock
Refactor the Kubernetes mock to move more code into shared internal
methods, most notably publication of internal events that is used
for the watch support. Add support for a callback after object
creation that can be used for things like creating the default
service account for a namespace when the namespace is created.