Switch to new syntax for generics
Use the new syntax added in Python 3.12 for generics. This results
in better typing in mypy, which required a few added casts because
the typing of metrics publishers is dynamic and somewhat strange.

Sphinx unfortunately has bugs generating documentation for generics
when `from __future__ import annotations` is present, so remove that
line from several modules so that the documentation will generate
properly.

autodoc_pydantic is unfortunately very buggy when used with
`Annotated`, which required several additional ignore rules and
probably results in poor documentation for some Pydantic models.