Add asyncmy and connect_args support for databases
The Qserv Kafka bridge needs to speak MySQL rather than PostgreSQL
to the Qserv frontend. Add support for rewriting `mysql` DSNs to
`mysql+asyncmy`, similar to the existing rewrite of `postgresql` DSNs,
to `create_async_engine`, and add support for a `connect_args`
parameter to pass arguments to the underlying database driver. The
latter is needed to configure TLS parameters with asyncmy.