Allow configuration of database pool size
Pass the `max_overflow`, `pool_size`, and `pool_timeout` arguments
down to the engine creation. Add all of the `create_database_engine`
to the `initialize` method of `DatabaseSessionDependency` as well,
and note them in the documentation.

This change still takes the approach of declaring each parameter
separately rather than passing all unknown parameters through to
SQLAlchemy. This results in better type checking at the cost of
having to change Safir for each new supported parameter. The code
is constructed to not pass a parameter, even with a value of `None`,
if that parameter was not set, mostly out of paranoia that some
engine parameter may give a special meaning to `None`.