Allow passing Kafka clients to make_manager
Add an optional `kafka_clients` argument to the `make_manager`
method of `safir.metrics.BaseMetricsConfiguration` to simplify the
integration of metrics into applications that want to reuse an
existing Kafka broker.

The previous recommended approach required manual construction of an
`EventManager`, which bypassed all of the configuration-based automatic
selection of no-op and mock event managers. Selection of those manager
variants is desirable even in an application with an existing Kafka
broker, since it may still want to test metrics or disable metrics
reporting.

Leave in the manual `KafkaEventManager` construction directions
just in case, since this is currently is the only way to provide a
pre-existing schema manager. Fix a few bugs in that documentation.