Always manage the admin client in EventManager
Applications, even if they use Kafka, will probably not need an
admin client, so having to create one and manage it in the case
where the application wants to reuse an existing Kafka broker for
an `EventManager` is awkward and undesirable. Undo the support
for passing an admin client into `make_manager` and only allow
passing in a Kafka broker, and change `EventManager` to
unconditionally manage the admin client.

Replace the `manage_kafka` parameter to `KafkaEventManager` with
`manage_kafka_broker` to reflect that it only controls management
of the Kafka broker, not the Kafka admin client.