metrics - don't blow up the app when app metrics blows up
If the underlying app metrics infrastructure is degraded, like if Kafka or the Schema Registry are not available, the Safir app metrics code now tries very hard to not crash or your instrumented application.

Instead of of raising exceptions, it will log error-level messages and put itself into an error state. Once in this error state, it will not even try to interact with any underlying infrastructure, which means it will not even try to send metrics.
It will instead only log error messages.

You will have to restart your app after the underlying infrastructure is fixed to start sending metrics again.