Revise documentation for current templates
There are some steps users should take after creating a new project
via Squarebot, so devote the main tutorial page to that and move
the manual process into a separate page with some duplication of
steps. Update both for the current templates (uv, Ruff instead of
Black, new dependency management method).

Update the README files to point to Phalanx in general and list the
Rubin Science Platform and Roundtable as separate cases. Drop the
link to the old Roundtable site, since that's no longer useful to
users.

Drop the list of Safir features from the README. It's out of date
and largely duplicates the documentation, which is already linked
from the README.