Avoid get_container_host_ip for schema registry
The latest testcontainers no longer supports `get_container_host_ip`
before the container has started, which broke the Safir support for
standing up a Confluent Schema Registry test container. Switch to
using `get_docker_client().host()`, which is what the previous
method used except in the case of unusual network connection modes.
This runs the risk that the other connection modes will not work,
but hopefully that won't be a problem for our use cases.