Redo the UWS pattern to call Wobbly in the worker
Originally, the UWS pattern stored job information directly in a
PostgreSQL database specific to the service and therefore needed a
separate database worker to store status and results because layering
the SQLAlchemy models on top of a stack container was likely to fail.
Subsequently, we switched to the Wobbly web service to store job
results, which can be accessed with much lighter-weight dependencies.
However, we kept the database worker to avoid major refactoring.

This is that delayed major refactoring. Eliminate the separate database
worker in favor of having the wrapper around the backend job function
update the job status and store results directly. This eliminates all
of the setup for the separate UWS worker and the code to force
serialization of tracebacks in exceptions so that they could be
serialized by arq in job results. Job workers no longer need to know
anything about the arq queue.

The (temporary) drawbacks are that safir.arq.uws gains a separate
Wobbly client from the proper Pydantic-based client in safir.uws, and
we temporarily lose Slack notifications of failures in the backend
worker because that would require a Safir dependency. Hopefully both
can be worked around in the future.