Convert to uv workspaces
Move the `safir` subdirectory to the top level and use uv workspaces
to manage safir-arq and safir-logging. Commit the top-level `uv.lock`
file following best-practices recommendations from the uv developers
for library packages. This freezes the dependencies for development
but has no effect on the dependencies of the uploaded library packages.

Adopt nox-uv and use its capabilities alongside workspaces to simplify
`noxfile.py`. nox-uv handles installing extras and dependency groups
directly, so there is no longer a need to handle that manually in the
`noxfile.py`.

For now, let the uv version float in GitHub Actions. This runs a small
risk that new uv releases will break tests in ways that are unrelated
to a given change, but we've had good luck with new uv versions so far.
If this becomes a problem in practice, we can adopt a more complicated
dependency management system with a script to update this version,
similar to what we use for service packages.