Move Kafka to extra, add import guards
Move the Kafka dependencies for safir.kafka and safir.metrics to a
new `kafka` extra so that other libraries can depend on Safir without
pulling in the Kafka library dependencies.

Add import guards to all the portions of Safir that require dependency
extras with a custom error message telling the user what extra is
required.

Add `from __future__ import annotations` to a few files that were
missing it and `__all__` to one file that didn't already have it.

Remove a mention of "pip dependencies" from the safir.database
documentation, since that's not entirely accurate these days. Mark a
few match strings as raw strings as required by Ruff 0.13.0.