Fix Starlette deprecation warning
Starlette deprecated the `HTTP_422_UNPROCESSABLE_ENTITY` constant and
changed it to `HTTP_422_UNPROCESSABLE_CONTENT` in [this
PR](https://github.com/Kludex/starlette/pull/2939). We could either
change this constant in our code, or just use the status code int
directly.

This PR uses status code directly here rather than the new constant in
to avoid having to increase the lower bound on starlette. This is the
choice that the FastAPI folks made in [this
PR](https://github.com/fastapi/fastapi/pull/14077/files).